Move some modules from flavours/glitch/utils/ back to flavours/glitch/features/ui/util/
