+++ app/javascript/flavours/glitch/features/compose/components/options.js
@@ -16,7 +16,7 @@ import LanguageDropdown from '../containers/language_dropdown_container';
 import ImmutablePureComponent from 'react-immutable-pure-component';
 
 //  Utils.
-import Motion from 'flavours/glitch/utils/optional_motion';
+import Motion from '../../ui/util/optional_motion';
 import { pollLimits } from 'flavours/glitch/initial_state';
 
 //  Messages.
