+++ app/javascript/flavours/glitch/features/ui/util/fullscreen.js
