+++ app/javascript/flavours/glitch/features/ui/util/get_rect_from_entry.js
