+++ app/javascript/flavours/glitch/features/ui/util/intersection_observer_wrapper.js
