+++ app/javascript/flavours/glitch/features/ui/util/optional_motion.js
