+++ app/javascript/flavours/glitch/features/ui/util/reduced_motion.js
