+++ app/javascript/flavours/glitch/features/ui/util/schedule_idle_task.js
